requirements.txt
aiocache==0.11.1
aiomysql==0.0.20
dataclasses==0.7
jieba_fast==0.53
nonebot==1.7.0
requests_toolbelt==0.9.1
